Use DecimalField for value in SensorValue and add aggregations by day

+++ server/models.py
@@ -79,7 +79,7 @@ class Sensor(models.Model):
 
 class SensorValue(models.Model):
     sensor = models.ForeignKey('Sensor')
-    value = models.FloatField()
+    value = models.DecimalField(max_digits=6, decimal_places=2)
     timestamp = models.DateTimeField(auto_now=False)
 
     def __unicode__(self):
